Compressed changes of bug 19329:  Reference to DataNodes is not removed when interactor is removed
update current selection to nullptr if nothing is selected

else a reference to a potentially deleted pointset is held, preventing
it from being destroyed

check for nullptr in DataNodeChange

to prevent crash when updating property lists

Dispatcher will also remove all DataInteractor with no valid DataNode

add a function to deal with deletion of DataNodes to DataInteractor

this was needed to have a way of removing a DataNode without triggering
modified request on the removed DataNode (as the corresponding DataInteracor
would be updated)
this function will be called when from the Dispatcher when a DataNode is removed.

extent test to check reference counts of Interactors and DataNodes

save paths in assemblycollection, so the datanode pointers are not needed.

make members private

Conflicts:
	Modules/Core/include/mitkVtkPropRenderer.h
	Modules/Core/src/Rendering/mitkVtkPropRenderer.cpp
	Modules/Core/test/mitkDispatcherTest.cpp
	Modules/DataTypesExt/src/mitkAffineDataInteractor3D.cpp